requirements.txt
Pillow
scipy
numpy
rembg
opencv-python
opencv-contrib-python